Refactor out the use of the trait Build
Build has been deprecated in favour of using .sbt files in sbt version 0.13.12
