+++ project/BuildDescriptionSettings.scala
@@ -14,61 +14,10 @@
  * limitations under the License.
  */
 
-import sbt.Keys._
-import sbt._
-import sbtassembly.AssemblyKeys._
-import sbtassembly._
-import uk.gov.hmrc.SbtAutoBuildPlugin
-import uk.gov.hmrc.DefaultBuildSettings._
-import uk.gov.hmrc.versioning.SbtGitVersioning
-
-
-object HmrcBuild extends Build {
-
-  val appName = "releaser"
-
-  val libraries = Seq(
-    "com.typesafe.play" %% "play-ws" % "2.3.8",
-    "com.jsuereth" %% "scala-arm" % "1.4",
-    "commons-io" % "commons-io" % "2.4",
-    "com.github.scopt" %% "scopt" % "3.3.0",
-    "org.apache.commons" % "commons-compress" % "1.10",
-    "org.scalatest" %% "scalatest" % "3.0.0" % "test",
-    "org.pegdown" % "pegdown" % "1.6.0" % "test",
-    "org.mockito" % "mockito-all" % "1.9.5" % "test"
-  )
-
-  lazy val releaser = Project(appName, file("."))
-    .enablePlugins(SbtAutoBuildPlugin, SbtGitVersioning)
-    .settings(
-      scalaVersion := "2.11.8",
-      libraryDependencies ++= libraries,
-      resolvers += Resolver.typesafeRepo("releases"),
-      BuildDescriptionSettings(),
-      AssemblySettings(),
-      addArtifact(artifact in (Compile, assembly), assembly)
-    )
-}
-
-object AssemblySettings{
-  def apply()= Seq(
-    assemblyJarName in assembly := "releaser.jar",
-    assemblyMergeStrategy in assembly := {
-      case PathList("org", "apache", "commons", "logging", xs@_*) => MergeStrategy.first
-      case PathList("play", "core", "server", xs@_*) => MergeStrategy.first
-      case x =>
-        val oldStrategy = (assemblyMergeStrategy in assembly).value
-        oldStrategy(x)
-    },
-    artifact in(Compile, assembly) := {
-      val art = (artifact in(Compile, assembly)).value
-      art.copy(`classifier` = Some("assembly"))
-    }
-  )
-}
-
 
 object BuildDescriptionSettings {
+
+  import sbt.Keys._
 
   def apply() =
     pomExtra := <url>https://www.gov.uk/government/organisations/hm-revenue-customs</url>
